requirements.txt
# Discord Trivia Bot Dependencies
# Python 3.14+ required

# Core Discord Bot
discord.py>=2.3.0

# Configuration Management
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code Quality (Development)
flake8>=6.0.0
black>=23.0.0
mypy>=1.0.0
